debug

+++ src/optical_flow/modules/pwcnet.py
@@ -126,7 +126,7 @@ class PWCNet(nn.Module):
         yy = torch.arange(0, H).view(-1, 1).repeat(1, W)
         xx = xx.view(1, 1, H, W).repeat(B, 1, 1, 1)
         yy = yy.view(1, 1, H, W).repeat(B, 1, 1, 1)
-        grid = torch.cat((xx, yy), 1).float()
+        grid = torch.cat((xx, yy), 1).float().type_as(flo)
 
         vgrid = Variable(grid) + flo
